+++ org/wikijava/sound/playWave/PlayWaveException.java
@@ -1,0 +1,20 @@
+package org.wikijava.sound.playWave;
+
+/**
+ * @author Giulio
+ */
+public class PlayWaveException extends Exception {
+
+    public PlayWaveException(String message) {
+	super(message);
+    }
+
+    public PlayWaveException(Throwable cause) {
+	super(cause);
+    }
+
+    public PlayWaveException(String message, Throwable cause) {
+	super(message, cause);
+    }
+
+}
